# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for CodeMentor AI
 * OpenAPI spec version: 0.1.0
 */
import type { AttemptInputLanguage } from './attemptInputLanguage';
import type { AttemptInputStatus } from './attemptInputStatus';

export interface AttemptInput {
  /** @nullable */
  problem_id?: string | null;
  language: AttemptInputLanguage;
  /** @maxLength 50000 */
  code: string;
  /**
     * @maxLength 10000
     * @nullable
     */
  output?: string | null;
  status: AttemptInputStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  hints_used?: number;
}